pyproject.toml
[project]
name = "the-fool"
version = "0.1.0"
description = "A discord bot"
authors = [
    { name = "Veirt", email = "veirt@duck.com" }
]
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.1",
    "aiohttp[speedups]>=3.9.3",
    "qbittorrent-api>=2024.1.58",
    "g4f>=0.2.1.0",
    "aria2p>=0.12.0",
    "bardapi @ git+https://github.com/veirt/Bard-API",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
bot = "the_fool:__main__"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
dev = "env DEV=True rye run bot"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/the_fool"]

[tool.pyright]
venvPath = "."
venv = ".venv"
